+++ modules/analysis/rosbag-image-pipeline-sfn/requirements-dev.txt
@@ -20,11 +20,15 @@ black==23.12.1
     # via -r requirements-dev.in
 boto3==1.34.28
     # via aws-sam-translator
+boto3-stubs==1.34.35
+    # via -r requirements-dev.in
 botocore==1.34.28
     # via
     #   awscli
     #   boto3
     #   s3transfer
+botocore-stubs==1.34.35
+    # via boto3-stubs
 build==1.0.3
     # via
     #   check-manifest
@@ -81,6 +85,10 @@ mccabe==0.7.0
 mpmath==1.3.0
     # via sympy
 mypy==1.8.0
+    # via -r requirements-dev.in
+mypy-boto3-dynamodb==1.34.34
+    # via -r requirements-dev.in
+mypy-boto3-s3==1.34.14
     # via -r requirements-dev.in
 mypy-extensions==1.0.0
     # via
@@ -162,14 +170,21 @@ sympy==1.12
     # via cfn-lint
 trove-classifiers==2024.1.8
     # via pyroma
+types-awscrt==0.20.3
+    # via botocore-stubs
 types-pyyaml==6.0.12.12
     # via -r requirements-dev.in
+types-s3transfer==0.10.0
+    # via boto3-stubs
 types-setuptools==69.0.0.20240125
     # via -r requirements-dev.in
 typing-extensions==4.9.0
     # via
     #   aws-sam-translator
+    #   boto3-stubs
     #   mypy
+    #   mypy-boto3-dynamodb
+    #   mypy-boto3-s3
     #   pydantic
     #   pydantic-core
 urllib3==2.0.7
